chore(ui): move MM components & store to `features/`
Somehow they had ended up in `features/ui/tabs` which isn't right
