+++ invokeai/frontend/web/src/features/modelManager/store/modelManagerSlice.ts
