+++ invokeai/frontend/web/src/features/modelManager/subpanels/AddModelsPanel/ScanModels.tsx
