+++ invokeai/frontend/web/src/features/modelManager/subpanels/AddModelsPanel/util.ts
