+++ invokeai/frontend/web/src/features/modelManager/subpanels/ImportModelsPanel.tsx
