+++ invokeai/frontend/web/src/features/modelManager/subpanels/MergeModelsPanel.tsx
