+++ invokeai/frontend/web/src/features/modelManager/subpanels/ModelManagerPanel.tsx
