+++ invokeai/frontend/web/src/features/modelManager/subpanels/ModelManagerPanel/ModelList.tsx
