+++ invokeai/frontend/web/src/features/modelManager/subpanels/ModelManagerSettingsPanel.tsx
