+++ invokeai/frontend/web/src/features/modelManager/subpanels/shared/BaseModelSelect.tsx
