+++ invokeai/frontend/web/src/features/modelManager/subpanels/shared/ModelVariantSelect.tsx
